+++ protected/humhub/views/htmlRedirect.php
@@ -1,22 +1,21 @@
-<script <?= \humhub\helpers\Html::nonce() ?>>
+<?php
 
+use humhub\helpers\Html;
+use yii\helpers\Json;
+
+/* @var $url string */
+
+// Encode once for a JavaScript string-literal context (this already includes the quotes).
+$jsUrl = Json::htmlEncode($url);
+// Remove test.php#xy (#xy) part, then encode the trimmed value as well.
+$jsUrlWithoutHash = Json::htmlEncode(explode('#', (string) $url)[0]);
+?>
+<script <?= Html::nonce() ?>>
     // If current URL == New Url (absolute || relative) then only Refresh
-    if (window.location.pathname + window.location.search + window.location.hash == '<?php echo $url; ?>' || '<?php echo $url; ?>' == window.location.href) {
-
-            <?php
-            //echo "window.location.reload();\n"; // Drops warning on Posts
-            // Remove test.php#xy  (#xy) part
-            $temp = explode("#", (string) $url);
-            $url = $temp[0];
-            ?>
-
-        if (window.location.search == '') {
-            window.location.href = '<?php echo $url; ?>';
-        } else {
-            window.location.href = '<?php echo $url; ?>';
-        }
-
+    if (window.location.pathname + window.location.search + window.location.hash == <?= $jsUrl ?>
+        || window.location.href == <?= $jsUrl ?>) {
+        window.location.href = <?= $jsUrlWithoutHash ?>;
     } else {
-        window.location.href = '<?php echo $url; ?>';
+        window.location.href = <?= $jsUrlWithoutHash ?>;
     }
 </script>